Refactor, add crossentropy and MNIST
Currently the value of input and output of function is always vector.
There is no need to make such a abstract, so I changed it into vector.
I intruduced a MNIST example, so I need a better loss function,
crossentropy. Since ./data is required in mnist.rs library, I moved
example outputs towards it and add it into .gitignore.